+++ src/main/java/com/ureca/uble/domain/users/repository/UsageHistoryRepository.java
@@ -1,9 +1,7 @@
-package com.ureca.uble.domain.usageHistory.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
+package com.ureca.uble.domain.users.repository;
 
 import com.ureca.uble.entity.UsageHistory;
+import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UsageHistoryRepository extends JpaRepository<UsageHistory, Long>, CustomUsageHistoryRepository{
-
 }
